feat: canonical basePath restructure — Launchpad owns root, SA owns /stock-signal/*
Per auth.md §URL architecture:
- Launchpad gains basePath-less config, deploys to S3 root (owns /, /sign-in/,
  /signed-out/, /launchpad/callback/)
- Stock Analyser gains basePath: /stock-signal, deploys to S3 stock-signal/ prefix
- Budget Tracker unchanged

CDK (deploy before apps):
- NetworkStack: replace /launchpad/* behavior with /stock-signal/* behavior
- AuthStack: logoutUrls updated to /signed-out/ for all three clients;
  Launchpad localhost callbackUrl corrected from port 3000 to 3001

SA route restructure:
- app/stock-signal/page.tsx → app/page.tsx (root)
- app/stock-signal/callback/page.tsx → app/callback/page.tsx
- Delete app/page.tsx (redirect stub), app/sign-in/, app/launchpad/
- handleGoToLaunchpad: router.push(/launchpad) → window.location.assign(/)

Launchpad route restructure:
- Remove basePath: /launchpad from next.config.mjs
- app/callback/page.tsx → app/launchpad/callback/page.tsx (preserves CDK-registered URL)
- Callback redirect: /launchpad/ → /; error link: /launchpad/sign-in/ → /sign-in/
- NEXT_PUBLIC_STOCK_URL default: localhost:3000 → localhost:3000/stock-signal/

Config fallback URLs (SA + BT): /launchpad/sign-in/ → /sign-in/; /launchpad/signed-out/ → /signed-out/
Workflow env vars: SIGNIN_URL, SIGNOUT_URL, STOCK_URL updated across all three workflows
Workflow S3 targets: SA syncs to stock-signal/ prefix; Launchpad syncs to root with excludes

auth-client (cognito-auth.ts): cross-client stale token fix — after UserAlreadyAuthenticatedException,
verify tokens belong to this client via getCurrentUser(); if not, clearCognitoStorage() and
retry signInWithRedirect. Replaces amplifySignOut() with clearCognitoStorage() to avoid
destroying tokens for the wrong client.

SA auth-guard + BT budget-tracker-app: restore authService.signInWithRedirect() for
unauthenticated state (reverting PR #218 redirect-to-launchpad approach); both apps must
initiate their own OAuth to consume the Cognito SSO session cookie per auth.md §7.

Docs: inventory.md CloudFront behaviors table updated to reflect completed restructure;
SA CLAUDE.md Known constraints updated.
