+++ .github/workflows/deploy-budget-tracker.yml
@@ -47,8 +47,8 @@ jobs:
       NEXT_PUBLIC_APP_URL: ${{ vars.NEXT_PUBLIC_APP_URL }}
       NEXT_PUBLIC_CALLBACK_URL: ${{ vars.NEXT_PUBLIC_APP_URL }}/budget-tracker/callback
       NEXT_PUBLIC_LAUNCHPAD_URL: ${{ vars.NEXT_PUBLIC_LAUNCHPAD_URL }}
-      NEXT_PUBLIC_SIGNIN_URL: ${{ vars.NEXT_PUBLIC_APP_URL }}/launchpad/sign-in/
-      NEXT_PUBLIC_SIGNOUT_URL: ${{ vars.NEXT_PUBLIC_APP_URL }}/launchpad/signed-out/
+      NEXT_PUBLIC_SIGNIN_URL: ${{ vars.NEXT_PUBLIC_APP_URL }}/sign-in/
+      NEXT_PUBLIC_SIGNOUT_URL: ${{ vars.NEXT_PUBLIC_APP_URL }}/signed-out/
       NEXT_PUBLIC_COMMIT_HASH: ${{ github.sha }}
     steps:
       - uses: actions/checkout@v4
@@ -112,8 +112,8 @@ jobs:
       NEXT_PUBLIC_APP_URL: ${{ vars.NEXT_PUBLIC_APP_URL }}
       NEXT_PUBLIC_CALLBACK_URL: ${{ vars.NEXT_PUBLIC_APP_URL }}/budget-tracker/callback
       NEXT_PUBLIC_LAUNCHPAD_URL: ${{ vars.NEXT_PUBLIC_LAUNCHPAD_URL }}
-      NEXT_PUBLIC_SIGNIN_URL: ${{ vars.NEXT_PUBLIC_APP_URL }}/launchpad/sign-in/
-      NEXT_PUBLIC_SIGNOUT_URL: ${{ vars.NEXT_PUBLIC_APP_URL }}/launchpad/signed-out/
+      NEXT_PUBLIC_SIGNIN_URL: ${{ vars.NEXT_PUBLIC_APP_URL }}/sign-in/
+      NEXT_PUBLIC_SIGNOUT_URL: ${{ vars.NEXT_PUBLIC_APP_URL }}/signed-out/
       NEXT_PUBLIC_COMMIT_HASH: ${{ github.sha }}
     steps:
       - uses: actions/checkout@v4
