+++ apps/launchpad/app/page.tsx
@@ -6,7 +6,7 @@ import { Launchpad } from '@/components/launchpad/launchpad'
 import { useAuthStore } from '@/stores/auth/use-auth-store'
 import { getConfig } from '@/lib/config'
 
-const STOCK_SIGNAL_URL = process.env.NEXT_PUBLIC_STOCK_URL ?? 'http://localhost:3000'
+const STOCK_SIGNAL_URL = process.env.NEXT_PUBLIC_STOCK_URL ?? 'http://localhost:3000/stock-signal/'
 
 export default function Page() {
   const router = useRouter()
