--- apps/stock-analyser/app/sign-in/page.tsx
@@ -1,12 +1,0 @@
-'use client'
-
-import { useEffect } from 'react'
-import { getConfig } from '@/lib/config'
-
-// SA does not initiate OAuth. Redirect to Launchpad's canonical /sign-in.
-export default function SignInPage() {
-  useEffect(() => {
-    window.location.replace(getConfig().apps.signInUrl)
-  }, [])
-  return null
-}
